ROSA provider: add option to use default account roles prefix
Previously when creating a cluster, it would create a new set of account
roles for each cluster, then delete them when the cluster is destroyed.

This commit adds a new field to the create cluster options struct,
allowing the caller to use the default account roles prefix
"ManagedOpenShift-X.Y" over creating unique ones per cluster. By
default, this is disabled and needs to be enabled.

+++ pkg/openshift/rosa/cluster.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"strings"
 	"time"
 
 	"github.com/Masterminds/semver"
@@ -16,13 +17,16 @@ import (
 	"sigs.k8s.io/e2e-framework/klient/wait"
 )
 
+const defaultAccountRolesPrefix = "ManagedOpenShift"
+
 // CreateClusterOptions represents data used to create clusters
 type CreateClusterOptions struct {
-	FIPS            bool
-	HostedCP        bool
-	MultiAZ         bool
-	STS             bool
-	SkipHealthCheck bool
+	FIPS                         bool
+	HostedCP                     bool
+	MultiAZ                      bool
+	STS                          bool
+	SkipHealthCheck              bool
+	UseDefaultAccountRolesPrefix bool
 
 	HostPrefix int
 	Replicas   int
@@ -97,7 +101,12 @@ func (r *Provider) CreateCluster(ctx context.Context, options *CreateClusterOpti
 		}
 		majorMinor := fmt.Sprintf("%d.%d", version.Major(), version.Minor())
 
-		accountRoles, err := r.createAccountRoles(ctx, options.ClusterName, majorMinor, options.ChannelGroup)
+		accountRolesPrefix := options.ClusterName
+		if options.UseDefaultAccountRolesPrefix {
+			accountRolesPrefix = fmt.Sprintf("%s-%s", defaultAccountRolesPrefix, majorMinor)
+		}
+
+		accountRoles, err := r.createAccountRoles(ctx, accountRolesPrefix, majorMinor, options.ChannelGroup)
 		if err != nil {
 			return "", &clusterError{action: action, err: err}
 		}
@@ -228,9 +237,11 @@ func (r *Provider) DeleteCluster(ctx context.Context, options *DeleteClusterOpti
 	}
 
 	if options.STS {
-		err = r.deleteAccountRoles(ctx, options.ClusterName)
-		if err != nil {
-			return &clusterError{action: action, err: err}
+		if !strings.Contains(cluster.AWS().STS().RoleARN(), defaultAccountRolesPrefix) {
+			err = r.deleteAccountRoles(ctx, options.ClusterName)
+			if err != nil {
+				return &clusterError{action: action, err: err}
+			}
 		}
 	}
